+++ src/annbatch/sampler/__init__.py
@@ -3,7 +3,8 @@
 This module provides samplers optimized for chunk-based data access patterns.
 """
 
-from annbatch.sampler._sampler import ChunkSampler, Sampler
+from annbatch.sampler._chunk_sampler import ChunkSampler
+from annbatch.sampler.abc import Sampler
 
 __all__ = [
     "ChunkSampler",
