+++ src/annbatch/sampler/abc/__init__.py
@@ -1,0 +1,5 @@
+from annbatch.sampler.abc._sampler import Sampler
+
+__all__ = [
+    "Sampler",
+]
